Move config to own module.
